+++ FuturaeKit.xcframework/ios-arm64/FuturaeKit.framework/Headers/FuturaeKit.h
@@ -16,6 +16,9 @@ FOUNDATION_EXPORT const unsigned char FuturaeKitVersionString[];
 
 // In this header, you should import all the public headers of your framework using statements like #import <FuturaeKit/PublicHeader.h>
 #import <FuturaeKit/FTRClient.h>
+#import <FuturaeKit/FTRAccount.h>
+#import <FuturaeKit/FTRMigratableAccount.h>
+#import <FuturaeKit/FTRMigrationCheckData.h>
 #import <FuturaeKit/FTRTotp.h>
 #import <FuturaeKit/FTRConfig.h>
 #import <FuturaeKit/FTRKeychainConfig.h>